[CarCareManager] Update the way to pass the vehicle id from vehicle list to vehicle details screen. The SelectedVehicleViewModel now automatically updates the selected vehicle after detecting any change in the database regarding the active vehicle

+++ CarCareManager/composeApp/src/commonMain/kotlin/com/echcoding/carcaremanager/data/repository/VehicleRepositoryImpl.kt
@@ -27,8 +27,8 @@ class VehicleRepositoryImpl(
         vehicleDao.deleteVehicleById(id)
     }
 
-    override suspend fun getVehicleById(id: Int): Vehicle? {
-        return vehicleDao.getVehicleById(id.toLong())?.toVehicle()
+    override suspend fun getVehicleById(id: Long): Vehicle? {
+        return vehicleDao.getVehicleById(id)?.toVehicle()
     }
 
     override suspend fun getAllVehicles(): Flow<List<Vehicle>> {
